requirements.txt
streamlit==1.22.0
tensorflow==2.11.0
opencv-python-headless==4.9.0.80
pillow
